requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
requests==2.27.1
urllib3==1.26.9
python-dotenv==1.0.1
